ArgumentParserTypes: implement a workaround for EDG 6 assertion failure
While building CMake on EDG 603+-based compiler, such as LCC 1.27.x,
an internal assertion failure may be experienced with a message like
'assertion failed at: "il.c", line 15635 in i_copy_dynamic_init'.

This commit introduces a workaround that helps avoiding this.

Workaround is enabled only for LCC based on this EDG frontend,
it is not known if it is needed for any other compilers, since
there were none found based on EDG 603 and later and having this bug.