FindOpenCL: Fix version detection in presence of multiple versions
Previously our call to `check_symbol_exists` was passing an absolute
path to the `cl.h` header to be included.  However, this might not
include its other headers correctly.  We already set
`CMAKE_REQUIRED_INCLUDES` such that `CL/cl.h` can be included by
the check, so we do not need to include by absolute path.