FindMotif: Update documentation
This syncs the documentation output with similar find modules.

Additionally, setting the MOTIF_FOUND variable to 0 is not needed at the
beginning of the module, therefore removed. The
find_package_handle_standard_arg() always sets the PackageName_FOUND and
PACKAGENAME_FOUND variables based on the required variables, also on
consecutive runs.