Utilities/ast-grep: add rules to remove `CM_DBG` usage
An easy way to detect and remove debugging left over from development.
The rules are disabled by default so that LSP usage doesn't interfere
when they are wanted.

To actually use the rules, severity flags are necessary:

    ast-grep scan --error=rm-cmdbg-includes --error=rm-cmdbg-macros

This command will report on instances; passing `-U` as well will update
the source.

Tests are also not run by default; see
https://github.com/ast-grep/ast-grep/issues/2023 to track its progress.