FindDependency: Split up macro to allow for no_return variant
Automatic dependency export requires a version of find_dependency that
does not automatically return on a failure to discover a dependency.
This allows the caller to perform clean-up prior to unwinding the
include stack.

Additionally, by making the dependency hash itself a macro "variable",
the hash stack mechanism is no longer necessary.