ci: Enable -Wstrict-prototypes in tests for merge request pipelines
Enable it in Makefile generator jobs that run in merge requests.
In particular, this causes RunCMake tests to fail if they build
C code with a missing function prototype because the compiler's
stderr reaches RunCMake's checks.  Catch this before staging.

Skip enabling this in the Fedora+Makefiles job because Fedora's GCC
now defaults to C23 under which -Wstrict-prototypes does not warn.

Suggested-by: Matthew Woehlke <matthew.woehlke@kitware.com>