Help: Note that some packaging preset keywords don't work as expected
Due to how the default value of the CPACK_PACKAGE_FILE_NAME variable
is implemented, the packageName and packageVersion fields of packaging
presets don't affect the final package file name. They can still affect other
aspects of the package produced, depending on the package generator
used, leading to inconsistencies in the generated package. There is no
warning or error about this when producing the packages, and fixing the
implementation is non-trivial. Until such a fix can be made, add a note to
the docs to make users aware of this unexpected behavior.

Issue: #25248