Help: Reposition internal modules
This adds a new section in the modules index to move some internal
modules out of the main two lists - utility modules and find modules.

They are not intended to be used by projects.

Additionally, SquishTestScript is not technically deprecated module. It
is just internally used by FindSquish module and mistakenly documented
and rendered.

Issue: #26851